messaging_app/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework>=3.15,<4.0
django-filter>=24.3
drf-nested-routers>=0.93.5
drf-yasg